refactor(tui): fold the new charts into chart.go and split the big ones
Codacy's per-file analysis flagged the two new chart files for missing
package comments — it cannot see doc.go from inside braille.go — and
AreaChart and Calendar for running past the complexity and length
limits. The charts live in chart.go with the rest of them now, and the
oversized functions give up their inner loops to named helpers:
brailleHeights, brailleCell, monthLabels, calendarCell, heatLegend, and
columnCell for the ColumnChart my per-column colours had pushed to the
edge of the same limit.
